Load database with questions at time of rendering quizzes to make it run in local machine
