+++ frontend/src/Home.js
@@ -27,37 +27,10 @@ function Home() {
       title: 'Quiz three',
       caption: 'Korea coronavirus'
     },
-
   ]
 
   return (
     <div className="Home">
-      {/* <nav className="navbar navbar-expand-lg navbar-dark fixed-top" id="mainNav">
-    <div className="container">
-      <a className="navbar-brand js-scroll-trigger" href="#page-top"></a>
-      <button className="navbar-toggler navbar-toggler-right" type="button" data-toggle="collapse" data-target="#navbarResponsive" aria-controls="navbarResponsive" aria-expanded="false" aria-label="Toggle navigation">
-        Menu
-        <i className="fa fa-bars"></i>
-      </button>
-      <div className="collapse navbar-collapse" id="navbarResponsive">
-        <ul className="navbar-nav text-uppercase ml-auto">
-         
-          <li className="nav-item">
-            <a className="nav-link js-scroll-trigger" href="#portfolio">Quizzes</a>
-          </li>
-          {/* <li className="nav-item">
-            <a className="nav-link js-scroll-trigger" href="Login">Login</a>
-          </li>
-          <li className="nav-item">
-            <a className="nav-link js-scroll-trigger" href="SignUp">Create Account</a>
-          </li> */}
-          
-        {/* </ul>
-      </div>
-    </div>
-  </nav> */}
-
-
       <header className="masthead">
         <div className="container">
           <div className="intro-text">
