--- frontend/src/quizService/index.js
@@ -1,374 +1,0 @@
-const qBank = [
-  {
-    question:
-      "When did corona virus first occurred?   ",
-    answers: [" November 2019", " December 2019", " January  2020", " December 2020"],
-    correct: " December 2019",
-    questionId: "099099"
-  },
-  {
-    question:
-      'Where did Corona first start?',
-    answers: ["Wuhan, China", "Seoul, Korea", "Beijing , China", "Tokyo, Japan"],
-    correct: "Wuhan, China",
-    questionId: "183452"
-  },
-  {
-    question:
-      "Coronavirus incubation period?",
-    answers: ["1~3", "1~7", "1~24", "1~14"],
-    correct: "1~14",
-    questionId: "267908"
-  },
-  {
-    question: " When was the first death reported? ",
-    answers: [
-      "January 2, 2020",
-      "January 7, 2020",
-      "January 19, 2020",
-      "January 10, 2020"
-    ],
-    correct: "January 10, 2020",
-    questionId: "333247"
-  },
-  {
-    question: "What animal is the starting point of the coronavirus?",
-    answers: ["Dog", "Bat", "Mouse", "Chicken"],
-    correct: "Bat",
-    questionId: "496293"
-  },
-  {
-    question:
-      "Is it neccesary to wear a mask when we go out?",
-    answers: [
-      "yes",
-      "Maybe",
-      "No"
-    ],
-    correct: "yes",
-    questionId: "588909"
-  },
-  {
-    question:
-      'How long we have to wash our hands?',
-    answers: ["30 second ", "10 second ", "15 second ", "20 second "],
-    correct: "20 second ",
-    questionId: "648452"
-  },
-  {
-    question: "What is NOT a good immune system?",
-    answers: ["Coke", "carrot", "pumpkin", "kiwi"],
-    correct: "Coke",
-    questionId: "786649"
-  },
-  {
-    question:
-      "What is NOT a major symptom of infection?",
-    answers: ["Fever 37.5", "cough", "parsley", "vomit"],
-    correct: "vomit",
-    questionId: "839754"
-  },
-  {
-    question:
-      "what number of mask will protect ous?",
-    answers: [
-      "Normal mask",
-      "KF70",
-      "KF60",
-      "KF80"
-    ],
-    correct: "KF80",
-    questionId: "98390"
-  },
-  {
-    question: "When did the coronavirus first confirm in Korea?",
-    answers: ["January 20th", "January 5th","January 10th","January 9th"],
-    correct: "January 20th",
-    questionId: "1071006"
-  },
-  {
-    question: " What is the date of the first death in Korea because of the coronavirus?",
-    answers: ["February 17th", "February 13th", "February 20th", "February 11th"],
-    correct: "February 20th",
-    questionId: "1174154"
-  },
-  {
-    question: "Which place has the most confirmed virus now?",
-    answers: ["IGC", "Seoul", "Songdo", "Daegu"],
-    correct: "Daegu",
-    questionId: "1226535"
-  },
-  {
-    question: "About how many patients are infected with the virus in Korea?",
-    answers: ["8000 people. ", " 9000 people. ", "10000 people. ", "7000 people. "],
-    correct: "8000 people. ",
-    questionId: "1310938"
-  },
-  {
-    question: "There is a vaccine in Korea now?",
-    answers: ["Yes", "No"],
-    correct: "No",
-    questionId: "1436873"
-  },
-  // {
-  //   question: "Which sign of the zodiac is represented by the Crab?",
-  //   answers: ["Cancer", "Libra", "Virgo", "Sagittarius"],
-  //   correct: "Cancer",
-  //   questionId: "1515110"
-  // },
-  // {
-  //   question: "On a dartboard, what number is directly opposite No. 1?",
-  //   answers: ["19", "20", "12", "15"],
-  //   correct: "19",
-  //   questionId: "1642728"
-  // },
-  // {
-  //   question:
-  //     "What does the 'S' stand for in the abbreviation SIM, as in SIM card? ",
-  //   answers: ["Subscriber", "Single", "Secure", "Solid"],
-  //   correct: "Subscriber",
-  //   questionId: "1747256"
-  // },
-  // {
-  //   question:
-  //     "What word represents the letter 'T' in the NATO phonetic alphabet?",
-  //   answers: ["Tango", "Target", "Taxi", "Turkey"],
-  //   correct: "Tango",
-  //   questionId: "1822532"
-  // },
-  // {
-  //   question: "Which American president appears on a one dollar bill?",
-  //   answers: [
-  //     "George Washington",
-  //     "Thomas Jefferson",
-  //     "Abraham Lincoln",
-  //     "Benjamin Franklin"
-  //   ],
-  //   correct: "George Washington",
-  //   questionId: "195075"
-  // },
-  // {
-  //   question: 'What is "dabbing"?',
-  //   answers: ["A dance", "A medical procedure", "A sport", "A language"],
-  //   correct: "A dance",
-  //   questionId: "2019778"
-  // },
-  // {
-  //   question: "What is the name of the Jewish New Year?",
-  //   answers: ["Rosh Hashanah", "Elul", "New Year", "Succoss"],
-  //   correct: "Rosh Hashanah",
-  //   questionId: "2134343"
-  // },
-  // {
-  //   question: "Which one of the following rhythm games was made by Harmonix?",
-  //   answers: [
-  //     "Rock Band",
-  //     "Meat Beat Mania",
-  //     "Guitar Hero Live",
-  //     "Dance Dance Revolution"
-  //   ],
-  //   correct: "Rock Band",
-  //   questionId: "2210799"
-  // },
-  // {
-  //   question:
-  //     "What type of animal was Harambe, who was shot after a child fell into it's enclosure at the Cincinnati Zoo?",
-  //   answers: ["Gorilla", "Tiger", "Panda", "Crocodile"],
-  //   correct: "Gorilla",
-  //   questionId: "2379831"
-  // },
-  // {
-  //   question: "Red Vines is a brand of what type of candy?",
-  //   answers: ["Licorice", "Lollipop", "Chocolate", "Bubblegum"],
-  //   correct: "Licorice",
-  //   questionId: "2426418"
-  // },
-  // {
-  //   question: "What is the nickname of the US state of California?",
-  //   answers: ["Golden State", "Sunshine State", "Bay State", "Treasure State"],
-  //   correct: "Golden State",
-  //   questionId: "2510086"
-  // },
-  // {
-  //   question: "What is on display in the Madame Tussaud's museum in London?",
-  //   answers: [
-  //     "Wax sculptures",
-  //     "Designer clothing",
-  //     "Unreleased film reels",
-  //     "Vintage cars"
-  //   ],
-  //   correct: "Wax sculptures",
-  //   questionId: "2685745"
-  // },
-  // {
-  //   question:
-  //     "What was the nickname given to the Hughes H-4 Hercules, a heavy transport flying boat which achieved flight in 1947?",
-  //   answers: ["Spruce Goose", "Noah's Ark", "Fat Man", "Trojan Horse"],
-  //   correct: "Spruce Goose",
-  //   questionId: "2796884"
-  // },
-  // {
-  //   question: "Which of these colours is NOT featured in the logo for Google?",
-  //   answers: ["Pink", "Yellow", "Blue", "Green"],
-  //   correct: "Pink",
-  //   questionId: "2838900"
-  // },
-  // {
-  //   question: 'What is the French word for "hat"?',
-  //   answers: ["Chapeau", "Bonnet", " Écharpe", " Casque"],
-  //   correct: "Chapeau",
-  //   questionId: "298060"
-  // },
-  // {
-  //   question: "Five dollars is worth how many nickles?",
-  //   answers: ["100", "50", "25", "69"],
-  //   correct: "100",
-  //   questionId: "3096579"
-  // },
-  // {
-  //   question: "Who is depicted on the US hundred dollar bill?",
-  //   answers: [
-  //     "Benjamin Franklin",
-  //     "George Washington",
-  //     "Abraham Lincoln",
-  //     "Thomas Jefferson"
-  //   ],
-  //   correct: "Benjamin Franklin",
-  //   questionId: "3182461"
-  // },
-  // {
-  //   question: "What do the letters in the GMT time zone stand for?",
-  //   answers: [
-  //     "Greenwich Mean Time",
-  //     "Global Meridian Time",
-  //     "General Median Time",
-  //     "Glasgow Man Time"
-  //   ],
-  //   correct: "Greenwich Mean Time",
-  //   questionId: "3239112"
-  // },
-  // {
-  //   question: "Which one of these is not a typical European sword design?",
-  //   answers: ["Scimitar", "Falchion", "Ulfberht", "Flamberge"],
-  //   correct: "Scimitar",
-  //   questionId: "3318503"
-  // },
-  // {
-  //   question:
-  //     'According to Sherlock Holmes, "If you eliminate the impossible, whatever remains, however improbable, must be the..."',
-  //   answers: ["Truth", "Answer", "Cause", "Source"],
-  //   correct: "Truth",
-  //   questionId: "3410327"
-  // },
-  // {
-  //   question: "What is the name of Poland in Polish?",
-  //   answers: ["Polska", "Pupcia", "Polszka", "Póland"],
-  //   correct: "Polska",
-  //   questionId: "3542596"
-  // },
-  // {
-  //   question: "The New York Times slogan is, “All the News That’s Fit to…”",
-  //   answers: ["Print", "Digest", "Look", "Read"],
-  //   correct: "Print",
-  //   questionId: "3667517"
-  // },
-  // {
-  //   question: "What do the letters of the fast food chain KFC stand for?",
-  //   answers: [
-  //     "Kentucky Fried Chicken",
-  //     "Kentucky Fresh Cheese",
-  //     "Kibbled Freaky Cow",
-  //     "Kiwi Food Cut"
-  //   ],
-  //   correct: "Kentucky Fried Chicken",
-  //   questionId: "3791672"
-  // },
-  // {
-  //   question: "Which restaurant's mascot is a clown?",
-  //   answers: ["McDonald's", "Whataburger", "Burger King", "Sonic"],
-  //   correct: "McDonald's",
-  //   questionId: "3893585"
-  // },
-  // {
-  //   question: 'What color is the "Ex" in FedEx Ground?',
-  //   answers: ["Green", "Red", "Light Blue", "Orange"],
-  //   correct: "Green",
-  //   questionId: "3913430"
-  // },
-  // {
-  //   question: "How tall is the Burj Khalifa?",
-  //   answers: ["2,722 ft", "2,717 ft", "2,546 ft", "3,024 ft"],
-  //   correct: "2,722 ft",
-  //   questionId: "4049121"
-  // },
-  // {
-  //   question:
-  //     "Which of the following card games revolves around numbers and basic math?",
-  //   answers: ["Uno", "Go Fish", "Twister", "Munchkin"],
-  //   correct: "Uno",
-  //   questionId: "4150746"
-  // },
-  // {
-  //   question:
-  //     "What machine element is located in the center of fidget spinners?",
-  //   answers: ["Bearings", "Axles", "Gears", "Belts"],
-  //   correct: "Bearings",
-  //   questionId: "4235063"
-  // },
-  // {
-  //   question: "Which sign of the zodiac comes between Virgo and Scorpio?",
-  //   answers: ["Libra", "Gemini", "Taurus", "Capricorn"],
-  //   correct: "Libra",
-  //   questionId: "4321002"
-  // },
-  // {
-  //   question: "Which of the following presidents is not on Mount Rushmore?",
-  //   answers: [
-  //     "John F. Kennedy",
-  //     "Theodore Roosevelt",
-  //     "Abraham Lincoln",
-  //     "Thomas Jefferson"
-  //   ],
-  //   correct: "John F. Kennedy",
-  //   questionId: "4442286"
-  // },
-  // {
-  //   question: "What is Tasmania?",
-  //   answers: [
-  //     "An Australian State",
-  //     "A flavor of Ben and Jerry's ice-cream",
-  //     "A Psychological Disorder",
-  //     "The Name of a Warner Brothers Cartoon Character"
-  //   ],
-  //   correct: "An Australian State",
-  //   questionId: "4564430"
-  // },
-  // {
-  //   question: "What company developed the vocaloid Hatsune Miku?",
-  //   answers: ["Crypton Future Media", "Sega", "Sony", "Yamaha Corporation"],
-  //   correct: "Crypton Future Media",
-  //   questionId: "4630606"
-  // },
-  // {
-  //   question:
-  //     "Which country, not including Japan, has the most people of japanese decent?",
-  //   answers: ["Brazil", "China", "South Korea", "United States of America"],
-  //   correct: "Brazil",
-  //   questionId: "4795960"
-  // },
-  // {
-  //   question: "Which candy is NOT made by Mars?",
-  //   answers: ["Almond Joy", "M&M's", "Twix", "Snickers"],
-  //   correct: "Almond Joy",
-  //   questionId: "4811162"
-  // },
-  // {
-  //   question: "In which fast food chain can you order a Jamocha Shake?",
-  //   answers: ["Arby's", "McDonald's", "Burger King", "Wendy's"],
-  //   correct: "Arby's",
-  //   questionId: "4982981"
-  // }
-];
-
-export default (n = 5) =>
-  Promise.resolve(qBank.sort(() => 0.5 - Math.random()).slice(0, n));
